Add primitive/include (#6379)
