+++ oneflow/core/kernel/boxing_zeros_kernel.cpp
@@ -15,7 +15,7 @@ limitations under the License.
 */
 #include "oneflow/core/kernel/kernel.h"
 #include "oneflow/core/kernel/kernel_context.h"
-#include "oneflow/core/primitive/memset.h"
+#include "oneflow/core/primitive/include/memset.h"
 
 namespace oneflow {
