+++ oneflow/core/primitive/cuda/memcpy.cpp
@@ -15,7 +15,7 @@ limitations under the License.
 */
 #ifdef WITH_CUDA
 
-#include "oneflow/core/primitive/memcpy.h"
+#include "oneflow/core/primitive/include/memcpy.h"
 #include "oneflow/core/stream/cuda_stream_context.h"
 #include "oneflow/core/primitive/cuda/cuda_graph_support.h"
 #include <cuda_runtime.h>
