+++ oneflow/core/primitive/cuda/memset.cpp
@@ -15,7 +15,7 @@ limitations under the License.
 */
 #ifdef WITH_CUDA
 
-#include "oneflow/core/primitive/memset.h"
+#include "oneflow/core/primitive/include/memset.h"
 #include "oneflow/core/stream/cuda_stream_context.h"
 #include "oneflow/core/primitive/cuda/cuda_graph_support.h"
 #include <cuda_runtime.h>
