+++ oneflow/core/primitive/include/add.h
@@ -13,35 +13,36 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_PRIMITIVE_MEMSET_H_
-#define ONEFLOW_CORE_PRIMITIVE_MEMSET_H_
+#ifndef ONEFLOW_CORE_PRIMITIVE_ADD_H_
+#define ONEFLOW_CORE_PRIMITIVE_ADD_H_
 
-#include "oneflow/core/primitive/primitive.h"
+#include "oneflow/core/primitive/include/primitive.h"
 
 namespace oneflow {
 
 namespace primitive {
 
-class Memset : public Primitive {
+class Add : public Primitive {
  public:
-  OF_DISALLOW_COPY_AND_MOVE(Memset);
-  Memset() = default;
-  ~Memset() override = default;
+  OF_DISALLOW_COPY_AND_MOVE(Add);
+  Add() = default;
+  ~Add() override = default;
 
-  virtual void Launch(StreamContext* stream_ctx, void* ptr, int value, size_t count) = 0;
+  virtual void Launch(StreamContext* stream_ctx, const void* const* srcs, size_t arity, void* dst,
+                      size_t count) = 0;
 };
 
-class MemsetFactory : public Factory<Memset> {
+class AddFactory : public Factory<Add> {
  public:
-  OF_DISALLOW_COPY_AND_MOVE(MemsetFactory);
-  MemsetFactory() = default;
-  ~MemsetFactory() override = default;
+  OF_DISALLOW_COPY_AND_MOVE(AddFactory);
+  AddFactory() = default;
+  ~AddFactory() override = default;
 
-  virtual std::unique_ptr<Memset> New() = 0;
+  virtual std::unique_ptr<Add> New(DataType data_type) = 0;
 };
 
 }  // namespace primitive
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_PRIMITIVE_MEMSET_H_
+#endif  // ONEFLOW_CORE_PRIMITIVE_ADD_H_
