+++ oneflow/core/primitive/include/cast.h
@@ -13,42 +13,35 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
-#define ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
+#ifndef ONEFLOW_CORE_PRIMITIVE_CAST_H_
+#define ONEFLOW_CORE_PRIMITIVE_CAST_H_
 
-#include "oneflow/core/primitive/primitive.h"
+#include "oneflow/core/primitive/include/primitive.h"
 
 namespace oneflow {
 
 namespace primitive {
 
-enum MemcpyKind {
-  kAuto = 0,
-  kHtoD,
-  kDtoH,
-  kDtoD,
+class Cast : public Primitive {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(Cast);
+  Cast() = default;
+  ~Cast() override = default;
+
+  virtual void Launch(StreamContext* stream_ctx, const void* from, void* to, size_t count) = 0;
 };
 
-class Memcpy : public Primitive {
+class CastFactory : public Factory<Cast> {
  public:
-  OF_DISALLOW_COPY_AND_MOVE(Memcpy);
-  Memcpy() = default;
-  ~Memcpy() override = default;
+  OF_DISALLOW_COPY_AND_MOVE(CastFactory);
+  CastFactory() = default;
+  ~CastFactory() override = default;
 
-  virtual void Launch(StreamContext* stream_ctx, void* dst, const void* src, size_t count) = 0;
-};
-
-class MemcpyFactory : public Factory<Memcpy> {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(MemcpyFactory);
-  MemcpyFactory() = default;
-  ~MemcpyFactory() override = default;
-
-  virtual std::unique_ptr<Memcpy> New(MemcpyKind kind) = 0;
+  virtual std::unique_ptr<Cast> New(DataType from, DataType to) = 0;
 };
 
 }  // namespace primitive
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
+#endif  // ONEFLOW_CORE_PRIMITIVE_CAST_H_
