+++ oneflow/core/primitive/include/memcpy.h
@@ -13,36 +13,42 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_PRIMITIVE_FILL_H_
-#define ONEFLOW_CORE_PRIMITIVE_FILL_H_
+#ifndef ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
+#define ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
 
-#include "oneflow/core/primitive/primitive.h"
-#include "oneflow/core/common/scalar.h"
+#include "oneflow/core/primitive/include/primitive.h"
 
 namespace oneflow {
 
 namespace primitive {
 
-class Fill : public Primitive {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(Fill);
-  Fill() = default;
-  ~Fill() override = default;
-
-  virtual void Launch(StreamContext* stream_ctx, void* dst, Scalar value, size_t count) = 0;
+enum MemcpyKind {
+  kAuto = 0,
+  kHtoD,
+  kDtoH,
+  kDtoD,
 };
 
-class FillFactory : public Factory<Fill> {
+class Memcpy : public Primitive {
  public:
-  OF_DISALLOW_COPY_AND_MOVE(FillFactory);
-  FillFactory() = default;
-  ~FillFactory() override = default;
+  OF_DISALLOW_COPY_AND_MOVE(Memcpy);
+  Memcpy() = default;
+  ~Memcpy() override = default;
 
-  virtual std::unique_ptr<Fill> New(DataType data_type) = 0;
+  virtual void Launch(StreamContext* stream_ctx, void* dst, const void* src, size_t count) = 0;
+};
+
+class MemcpyFactory : public Factory<Memcpy> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(MemcpyFactory);
+  MemcpyFactory() = default;
+  ~MemcpyFactory() override = default;
+
+  virtual std::unique_ptr<Memcpy> New(MemcpyKind kind) = 0;
 };
 
 }  // namespace primitive
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_PRIMITIVE_FILL_H_
+#endif  // ONEFLOW_CORE_PRIMITIVE_MEMCPY_H_
